+++ apps/api/v2/src/ee/schedules/schedules.module.ts
@@ -1,5 +1,6 @@
 import { SchedulesController } from "@/ee/schedules/controllers/schedules.controller";
 import { SchedulesRepository } from "@/ee/schedules/schedules.repository";
+import { ResponseService } from "@/ee/schedules/services/response/response.service";
 import { SchedulesService } from "@/ee/schedules/services/schedules.service";
 import { AvailabilitiesModule } from "@/modules/availabilities/availabilities.module";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
@@ -8,7 +9,7 @@ import { Module } from "@nestjs/common";
 
 @Module({
   imports: [PrismaModule, AvailabilitiesModule, UsersModule],
-  providers: [SchedulesRepository, SchedulesService],
+  providers: [SchedulesRepository, SchedulesService, ResponseService],
   controllers: [SchedulesController],
 })
 export class SchedulesModule {}
