+++ apps/api/v2/src/ee/schedules/services/response/zod/response.ts
